feat: add ability to subscribe to certain namespaces for message notifications
In order to not automtically lock the resources I've introduced a Lock
method in the interface to let the client decide if they just want
the notification or also lock the message for use.

I'd recommend not to use Subscribe + Dequeue at the same time. Pick one
or the other.